[Actor] Move actor files into data/ directory
