+++ src/globals/CMakeLists.txt
@@ -1,7 +1,5 @@
 add_library(
   mediaelch_globals OBJECT
-  Actor.cpp
-  ActorModel.cpp
   ComboDelegate.cpp
   Containers.cpp
   DownloadManager.cpp
